Refactored all JSON aspects of check messages into a new protocol/check package
